+++ db/migrate/20200102142041_add_created_by_to_payroll_runs.rb
@@ -1,0 +1,5 @@
+class AddCreatedByToPayrollRuns < ActiveRecord::Migration[6.0]
+  def change
+    add_reference :payroll_runs, :created_by, type: :uuid, foreign_key: {to_table: :dfe_sign_in_users}
+  end
+end